move TagKeyDeletionTask
